Add player and listener classes to utils.py
