+++ web/fastapi_ws.py
@@ -1,4 +1,3 @@
-from fastapi import FastAPI, WebSocket
 import uvicorn
 from openvoicechat.tts.tts_piper import Mouth_piper as Mouth
 from openvoicechat.llm.llm_gpt import Chatbot_gpt as Chatbot
@@ -7,72 +6,16 @@ from openvoicechat.llm.prompts import llama_sales
 from openvoicechat.utils import run_chat
 from dotenv import load_dotenv
 import os
-import librosa
 import threading
-import numpy as np
 import queue
-import time
 from fastapi import FastAPI, WebSocket, WebSocketDisconnect
 from fastapi.responses import FileResponse
 from fastapi.staticfiles import StaticFiles
-
+from openvoicechat.utils import Listener_ws, Player_ws
 
 app = FastAPI()
 
 app.mount("/static", StaticFiles(directory="static"))
-
-
-class Player_ws:
-    def __init__(self, q):
-        self.output_queue = q
-        self.playing = False
-
-    def play(self, audio_array, samplerate):
-        audio_array = audio_array / (1 << 15)
-        audio_array = audio_array.astype(np.float32)
-        audio_array = librosa.resample(y=audio_array, orig_sr=samplerate, target_sr=44100)
-        audio_array = audio_array.tobytes()
-        self.output_queue.put(audio_array)
-
-    def stop(self):
-        self.playing = False
-        self.output_queue.queue.clear()
-        self.output_queue.put('stop'.encode())
-
-    def wait(self):
-        time_to_wait = 0
-        # while not self.output_queue.empty():
-        #     time.sleep(0.1)
-        #     peek at the first element
-            # time_to_wait = len(self.output_queue.queue[0]) / (44100 * 4)
-        # print(time_to_wait)
-        # time.sleep(time_to_wait)
-        self.playing = False
-
-
-class Listener_ws:
-    def __init__(self, q):
-        self.input_queue = q
-        self.listening = False
-        self.CHUNK = 5945
-        self.RATE = 16_000
-
-    def read(self, x):
-        data = self.input_queue.get()
-        data = np.frombuffer(data, dtype=np.float32)
-        data = librosa.resample(y=data, orig_sr=44100, target_sr=16_000)
-        data = data * (1 << 15)
-        data = data.astype(np.int16)
-        data = data.tobytes()
-        return data
-
-    def close(self):
-        pass
-
-    def make_stream(self):
-        self.listening = True
-        self.input_queue.queue.clear()
-        return self
 
 
 @app.websocket("/ws")
@@ -91,11 +34,7 @@ async def websocket_endpoint(websocket: WebSocket):
     api_key = os.getenv('OPENAI_API_KEY')
     chatbot = Chatbot(sys_prompt=llama_sales,
                       api_key=api_key)
-    # run transcribe in thread
-    # threading.Thread(target=transcribe, args=(ear, listener)).start()
-    # threading.Thread(target=play_text, args=(mouth, player, 'Hello, my name is John.')).start()
     threading.Thread(target=run_chat, args=(mouth, ear, chatbot, True)).start()
-
 
     try:
         while True:
@@ -118,8 +57,11 @@ async def websocket_endpoint(websocket: WebSocket):
     finally:
         await websocket.close()
 
+
 @app.get("/")
 def read_root():
     return FileResponse('static/stream_audio.html')
+
+
 if __name__ == "__main__":
     uvicorn.run(app, host='0.0.0.0', port=8000)
